improve fetchBatch
