+++ packages/common/src/audius-query/schema.ts
@@ -1,4 +1,4 @@
-import { schema } from 'normalizr'
+import { Schema, schema } from 'normalizr'
 
 import { Kind } from '~/models/Kind'
 
@@ -30,7 +30,7 @@ export const userManagerSchema = new schema.Object({
   manager: userSchema
 })
 
-export const apiResponseSchema = new schema.Object({
+export const schemas = {
   currentWeb3User: userSchema,
   managedUsers: new schema.Array(managedUserSchema),
   user: userSchema,
@@ -40,4 +40,6 @@ export const apiResponseSchema = new schema.Object({
   users: new schema.Array(userSchema),
   tracks: new schema.Array(trackSchema),
   collections: new schema.Array(collectionSchema)
-})
+}
+
+export const apiResponseSchema = new schema.Object(schemas)
